DFE-507 Move table tool components into frontend app
