--- src/explore-education-statistics-frontend/src/modules/table-tool/components/CharacteristicsFilterForm.module.scss
@@ -1,6 +1,0 @@
-@import '~explore-education-statistics-common/src/styles/govuk-base';
-
-.year {
-  display: inline-block;
-  margin-right: govuk-spacing(6);
-}
