+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/FixedHeaderGroupedDataTable.module.scss
@@ -1,4 +1,4 @@
-@import '~explore-education-statistics-common/src/styles/govuk-base';
+@import '../../../../node_modules/explore-education-statistics-common/src/styles/govuk-base';
 
 $border-width: 4px;
