+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/FormCheckboxSelectedCount.module.scss
@@ -1,0 +1,6 @@
+@import '../../../../node_modules/explore-education-statistics-common/src/styles/govuk-base';
+
+.selectionCount {
+  font-size: 0.8rem;
+  margin-left: govuk-spacing(2);
+}
