+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/FormSortableList.module.scss
@@ -1,4 +1,4 @@
-@import '~explore-education-statistics-common/src/styles/govuk-base';
+@import '../../../../node_modules/explore-education-statistics-common/src/styles/govuk-base';
 
 .list {
   padding: govuk-spacing(2);
