+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/TableHeadersForm.tsx
@@ -1,12 +1,12 @@
 import Button from '@common/components/Button';
 import Details from '@common/components/Details';
 import TimePeriod from '@common/services/types/TimePeriod';
-import FormFieldSortableList from '@frontend/prototypes/table-tool/components/FormFieldSortableList';
-import { SortableOption } from '@frontend/prototypes/table-tool/components/FormSortableList';
+import FormFieldSortableList from '@frontend/modules/table-tool/components/FormFieldSortableList';
+import { SortableOption } from '@frontend/modules/table-tool/components/FormSortableList';
 import {
   FilterOption,
   IndicatorOption,
-} from '@frontend/prototypes/table-tool/components/meta/initialSpec';
+} from '@frontend/modules/table-tool/components/meta/initialSpec';
 import { Form, Formik } from 'formik';
 import React from 'react';
