+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/Wizard.module.scss
@@ -1,0 +1,9 @@
+@import '../../../../node_modules/explore-education-statistics-common/src/styles/govuk-base';
+
+//.container {
+//}
+
+.stepNav {
+  list-style: none;
+  padding-left: 0;
+}
