+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/WizardStepHeading.module.scss
@@ -1,4 +1,4 @@
-@import '~explore-education-statistics-common/src/styles/govuk-base';
+@import '../../../../node_modules/explore-education-statistics-common/src/styles/govuk-base';
 
 $circle-size: 40px;
 $circle-border-size: 4px;
