+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/meta/initialSpec.ts
