+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/utils/parseAcademicYear.ts
