--- src/explore-education-statistics-frontend/src/pages/prototypes/table-tool.tsx
@@ -1,3 +1,0 @@
-import Page from '@frontend/prototypes/table-tool/PrototypeTableToolPage';
-
-export default Page;
