--- src/explore-education-statistics-frontend/src/prototypes/table-tool/components/FormCheckboxSelectedCount.module.scss
@@ -1,6 +1,0 @@
-@import '~explore-education-statistics-common/src/styles/govuk-base';
-
-.selectionCount {
-  font-size: 0.8rem;
-  margin-left: govuk-spacing(2);
-}
