--- src/explore-education-statistics-frontend/src/prototypes/table-tool/components/Wizard.module.scss
@@ -1,9 +1,0 @@
-@import '~explore-education-statistics-common/src/styles/govuk-base';
-
-//.container {
-//}
-
-.stepNav {
-  list-style: none;
-  padding-left: 0;
-}
